Move more code to dataset downloader
